+++ src/Markers/MarkerDivIcon.ts
@@ -1,4 +1,4 @@
-import * as Schema from "../JSONSchema";
+import * as Schema from "../Interfaces/JSONSchema";
 import { DivIcon, Point, DomUtil } from "leaflet";
 
 export interface NamedDivIcon {
