+++ src/WSMap.ts
@@ -6,8 +6,8 @@ import { ControlDock } from "./Controls/ControlDock";
 import { ZoomControl } from "./Controls/ZoomControl";
 import { FilterControl } from "./Controls/FilterControl";
 import { LayersControl } from "./Controls/LayersControl";
-import { FilterCategory } from "./FilterCategory";
-import * as Schema from "./JSONSchema";
+import { FilterCategory } from "./Interfaces/FilterCategory";
+import * as Schema from "./Interfaces/JSONSchema";
 
 export interface WSMapOptions extends MapOptions {
   mapSizePixels: number;
